Be lazier: /services no longer includes command info.
To see the commands, services/index is not enough.
Instead, services/show does it, via a redirect to commands/index.
This enables seeing all services without unnecessarily
asking for too much detail.

/services is now sorted, disable by ?sort=0.

Removed caching the services in the session cookie:
the real service hash was too big for this.

commands/update:
  call /etc/init.d/FOO, /usr/sbin/rcFOO is nonstandard;
  always return exit+stdout+stderr