Make PackageKit mocking complete
This commit completes the mocking of PackageKit via D-Bus.

It introduces a PackageKitStub class and a PackageKitResult class.

PackageKitStub covers the complete D-Bus communication, PackageKitResult
the expected results from a function call (i.e. "GetUpdates").

The mocking is complete up to and including the socket level. The
results are sent to the socket and then processed through ruby-dbus,
making it completely transparent to the application.