Fixed a hang when PackageKit responded too fast (bnc#561578).
All signal handlers need to be in place before the call is made
otherwise the signal will be emitted together with the reply
(and ignored).

A test case is not possible with the current
test/packagekit_stub.rb,
because it cannot work before run is called.
But the bug strikes only when Finished arrives before run is called,
even earlier, before the handler is installed.