Report correct REST service user in NoPermissionException, plus testcase
This fixes bnc#601939 and adds a first testcase to lib/yast_service

Permission handling in rest-service has two sides. One is the
permissions of the current session owner and the other permissions of
the rest-service itself in starting DBus services - like YaST.

This test is about the latter and tries to report the correct user id,
the one the rest-service is running under.