fix passing checkbox value to controller
Reason to this change is that rails have new feature, that
unchecked value (last value for check_box helper) is evaluated
and if it is false, it doesn't generate unchecked value and it is not
passed at all to controller.

Previous behavior was to pass just value.to_s, so I manually change
values from true,false to "true,"false" that is always true and same as
previous to_s behavior.

There is two exceptions:
1) in time checkbox is used only for javascript, so it doesn't need
special values and I remove it to kept defaults
2) in kerberos I add special handling for use_dns as it doesn't match
dns_used value in model. Looks like relict from from server merge and
because it is used by javascript and model use name from ycp I don't
want to break it and add just translation.